+++ app/src/main/java/com/gowthamraj07/journeytracker/domain/Place.kt
@@ -1,8 +1,8 @@
 package com.gowthamraj07.journeytracker.domain
 
 data class Place(
-    val id: Int,
-    val tripId: Int,
+    val id: Long,
+    val tripId: Long,
     val latitude: Double,
     val longitude: Double,
     val imageUrl: String
